Address docker support issues
When running the script in Docker for Windows and when Plex is running natively on Windows, there was an issue where the Windows paths returned by Plex for the different media items could not be parsed natively inside the container (as it runs Linux) to find the proper asset folder for an item. I added some logic to determine whether a path obtained by the script from Plex is a Windows or Linux path so that it can be parsed correctly.

I also fine-tuned the docker-compose and Dockerfile files to map the right Kometa base folder to the container and to run the script in debug mode so that the container logs show the debug output for better troubleshooting.

And in the process of troubleshooting I found a bug where ever time you saved the config file from the web UI the process to set the TV show and movie libraries to the plex object would just append the new set of libraries to the existing one, and this would make the script iterate multiple times over the same libraries.