Minor fixes and improvements
Fixed minor cosmetic issue in plex_connector.py. Removed unnecessary properties from the KometaSaver class. Fixed some issues with proper artwork type handling in upload_processor.py (differentiate between poster and background for movies and collections) as well as allowing the saving of missing episode artwork when save_to_kometa or --kometa are set, since there's no need for the episode to actually be on Plex for the script to download it and save it to its asset folder so it can be applied as soon as the episode is downloaded and Kometa runs.